add time bar at top of page
Add a sticky timebar at top of page so you know which room is free when.
Uses flexboxes unlike the rest of the project. However, this change is
very modular (e.g. it won't break if you change how many timeslots are
displayed) and works well when resizing the browser window. This
modularity with flexboxes also allows for easy restyling.
Note this also works well on mobile.